comment and refactor
