+++ mypyc/irbuild/builder.py
@@ -35,12 +35,11 @@ from mypyc.ir.ops import (
     AssignmentTargetAttr, AssignmentTargetTuple, Environment, LoadInt, Value,
     Register, Op, Assign, Branch, Unreachable, TupleGet, GetAttr, SetAttr, LoadStatic,
     InitStatic, PrimitiveOp, OpDescription, NAMESPACE_MODULE, RaiseStandardError,
-    BinaryIntOp
 )
 from mypyc.ir.rtypes import (
     RType, RTuple, RInstance, int_rprimitive, dict_rprimitive,
     none_rprimitive, is_none_rprimitive, object_rprimitive, is_object_rprimitive,
-    str_rprimitive, c_pyssize_t_rprimitive, bool_rprimitive
+    str_rprimitive,
 )
 from mypyc.ir.func_ir import FuncIR, INVALID_FUNC_DEF
 from mypyc.ir.class_ir import ClassIR, NonExtClassInfo
@@ -48,7 +47,6 @@ from mypyc.primitives.registry import func_ops, CFunctionDescription, c_function
 from mypyc.primitives.list_ops import list_len_op, to_list, list_pop_last
 from mypyc.primitives.dict_ops import dict_get_item_op, dict_set_item_op
 from mypyc.primitives.generic_ops import py_setattr_op, iter_op, next_op
-from mypyc.primitives.int_ops import int_logical_op_mapping
 from mypyc.primitives.misc_ops import true_op, false_op, import_op
 from mypyc.crash import catch_errors
 from mypyc.options import CompilerOptions
@@ -242,31 +240,8 @@ class IRBuilder:
         return self.builder.environment
 
     # Helpers for IR building
-    def check_tagged_short_int(self, val: Value, line: int) -> Value:
-        int_tag = self.add(LoadInt(1, line, rtype=c_pyssize_t_rprimitive))
-        bitwise_and = self.binary_int_op(c_pyssize_t_rprimitive, val,
-                                         int_tag, BinaryIntOp.AND, line)
-        zero = self.add(LoadInt(0, line, rtype=c_pyssize_t_rprimitive))
-        check = self.binary_int_op(bool_rprimitive, bitwise_and, zero, BinaryIntOp.EQ, line)
-        return check
-
-    def binary_comparison_op(self, lhs: Value, rhs: Value, op: str, line: int) -> Value:
-        op_type, c_func_desc = int_logical_op_mapping[op]
-        result = self.alloc_temp(bool_rprimitive)
-        short_int_block, int_block, out = BasicBlock(), BasicBlock(), BasicBlock()
-        check = self.check_tagged_short_int(lhs, line)
-        branch = Branch(check, short_int_block, int_block, Branch.BOOL_EXPR)
-        branch.negated = False
-        self.add(branch)
-        self.activate_block(short_int_block)
-        eq = self.binary_int_op(bool_rprimitive, lhs, rhs, op_type, line)
-        self.assign(result, eq, line)
-        self.goto(out)
-        self.activate_block(int_block)
-        call = self.call_c(c_func_desc, [lhs, rhs], line)
-        self.assign(result, call, line)
-        self.goto_and_activate(out)
-        return result
+    def compare_tagged(self, lhs: Value, rhs: Value, op: str, line: int) -> Value:
+        return self.builder.compare_tagged(lhs, rhs, op, line)
 
     def add_to_non_ext_dict(self, non_ext: NonExtClassInfo,
                             key: str, val: Value, line: int) -> None:
